remove const bitcast folding [PR]
weak const won't have width. _bits_to_rand calls python bitcast directly
